add cancelItemUpload()
