+++ packages/arcgis-rest-portal/src/index.ts
@@ -9,7 +9,7 @@ export * from "./items/remove";
 export * from "./items/search";
 export * from "./items/update";
 export * from "./items/helpers";
-export * from "./items/commit";
+export * from "./items/upload";
 
 export * from "./groups/add-users";
 export * from "./groups/create";
